Refactor transaction metadata imports and remove unused metadata module
