+++ src/models/results/account_tx.rs
@@ -7,9 +7,8 @@ use serde_json::Value;
 use crate::models::requests::Marker;
 use crate::models::{XRPLModelException, XRPLModelResult};
 
-use super::{
-    exceptions::XRPLResultException, metadata::TransactionMetadata, XRPLResponse, XRPLResult,
-};
+use super::{exceptions::XRPLResultException, XRPLResponse, XRPLResult};
+use crate::models::transactions::metadata::TransactionMetadata;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(untagged)]
